// apps/user-web/src/.generated/api/model/apiErrorResponseDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * User API
 * Control Plane User API Service
 * OpenAPI spec version: 1.0.0
 */
import type { ApiErrorResponseDtoData } from './apiErrorResponseDtoData';
import type { ApiErrorResponseDtoDetails } from './apiErrorResponseDtoDetails';
import type { ApiErrorResponseDtoMeta } from './apiErrorResponseDtoMeta';

export interface ApiErrorResponseDto {
  success: boolean;
  statusCode: number;
  path: string;
  requestId: string;
  timestamp: string;
  message?: string;
  /** @nullable */
  data: ApiErrorResponseDtoData;
  meta?: ApiErrorResponseDtoMeta;
  errorCode: string;
  /**
     * Validation error details containing fields mapping
     * @nullable
     */
  details?: ApiErrorResponseDtoDetails;
}